[cosmos] Migrate @azure/cosmos to ESM/vitest

+++ sdk/cosmosdb/cosmos/test/public/functional/throughput.spec.ts
@@ -188,7 +188,9 @@ describe("testThroughputBucketForBulk", async () => {
   });
 
   afterAll(async () => {
-    await container.database.delete();
+    if (container) {
+      await container.database.delete();
+    }
   });
 
   it("test throughput bucketing in bulk", async () => {
